Restructure packaging for dotpay
COAND-497
